First Commit after initialization
it's been about a week that i've begun to learn C programming language in a deep dive style of learning using the book by k.n.king. after about a week i've noticed that this book has a lot of programs. most of the projects and exercises in this book have not been provided by a solution. rather a hnadful of them have been provided with a solid program written for it. anyways, this is purely educational. (is it really? TBH IDK either).